% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insidesp-package.R
\docType{package}
\name{insidesp-package}
\alias{insidesp}
\alias{insidesp-package}
\title{insidesp: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Michael D. Sumner \email{mdsumner@gmail.com}

Other contributors:
\itemize{
  \item Edzer Pebesma (wrote the code in sp) [contributor]
}

}
\keyword{internal}
